fix(domain/service/mapper): corrige erros de compilacao detectados no build
- Equipamento.java: corrige import AuditableEntity (config -> core.domain) e substitui @Where por @SQLRestriction
- CronogramaService.java: substitui ResourceNotFoundException por RecursoNaoEncontradoException
- AnexoService.java: corrige tipo Long -> Integer em buscarPorId() e deletar()
- AnexoController.java: corrige @PathVariable Long -> Integer para alinhar com Anexo.id
- AnexoMapper.java: remove imports duplicados
- Insumo.java: corrige @AllArgsConstructor para access = PRIVATE (forcando uso do builder pelo MapStruct)
- FaseCronograma.java: garante @Getter explicito na classe (ja tinha, mas garante compatibilidade)

+++ pruma/src/main/java/com/br/pruma/application/service/AnexoService.java
@@ -1,6 +1,5 @@
 package com.br.pruma.application.service;
 
-
 import com.br.pruma.application.dto.request.AnexoRequestDTO;
 import com.br.pruma.application.dto.response.AnexoResponseDTO;
 import com.br.pruma.application.mapper.AnexoMapper;
@@ -8,6 +7,7 @@ import com.br.pruma.config.RecursoNaoEncontradoException;
 import com.br.pruma.core.domain.Anexo;
 import com.br.pruma.core.repository.AnexoRepository;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -29,7 +29,7 @@ public class AnexoService {
                 .collect(Collectors.toList());
     }
 
-    public AnexoResponseDTO buscarPorId(Long id) {
+    public AnexoResponseDTO buscarPorId(Integer id) {
         Anexo anexo = repository.findById(id)
                 .orElseThrow(() -> new RecursoNaoEncontradoException("Anexo com ID " + id + " não encontrado."));
         return mapper.toResponseDTO(anexo);
@@ -41,11 +41,10 @@ public class AnexoService {
         return mapper.toResponseDTO(salvo);
     }
 
-    public void deletar(Long id) {
+    public void deletar(Integer id) {
         if (!repository.existsById(id)) {
             throw new RecursoNaoEncontradoException("Não é possível deletar. Anexo com ID " + id + " não existe.");
         }
         repository.deleteById(id);
     }
 }
-
